refac: 배틀 매치 화면 개발

+++ application/battle-application/src/main/kotlin/com/monglife/mongs/application/battle/vo/MatchVo.kt
@@ -31,7 +31,7 @@ data class MatchVo(
         val mongCode: String,
         val mongName: String,
         val name: String,
-        val hp: Double,
+        val hp: Float,
         val roundCode: MatchRoundCode,
         val isMe: Boolean,
     ) {
