+++ data/battle-data/src/main/kotlin/com/monglife/mongs/data/battle/persistence/dto/MatchEventDto.kt
@@ -21,7 +21,7 @@ data class MatchEventDto(
         val mongCode: String,
         val mongName: String,
         val name: String,
-        val hp: Double,
+        val hp: Float,
         val roundCode: MatchRoundCode,
     )
 }
